+++ cmd/agent-server/main.go
@@ -1,0 +1,73 @@
+package main
+
+import (
+	"context"
+	"database/sql"
+	"fmt"
+	"net"
+	"net/http"
+	"os"
+	"os/signal"
+	"syscall"
+	"time"
+
+	"github.com/aijustin/agentflow-go/internal/cmd/agentruntime"
+)
+
+func main() {
+	config, err := agentruntime.LoadConfigFromEnv()
+	if err != nil {
+		_, _ = fmt.Fprintln(os.Stderr, err)
+		os.Exit(1)
+	}
+	var db *sql.DB
+	defer func() {
+		if db != nil {
+			_ = db.Close()
+		}
+	}()
+	fw, err := agentruntime.NewFramework(config, os.Stderr)
+	if err != nil {
+		_, _ = fmt.Fprintln(os.Stderr, err)
+		os.Exit(1)
+	}
+	queue, err := agentruntime.NewQueue(config, &db)
+	if err != nil {
+		_, _ = fmt.Fprintln(os.Stderr, err)
+		os.Exit(1)
+	}
+	handler, err := agentruntime.NewProductionHandler(config, fw, queue)
+	if err != nil {
+		_, _ = fmt.Fprintln(os.Stderr, err)
+		os.Exit(1)
+	}
+	server := &http.Server{
+		Addr:              config.HTTPAddr,
+		Handler:           handler,
+		ReadHeaderTimeout: 5 * time.Second,
+	}
+	ctx, stop := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGTERM)
+	defer stop()
+	go func() {
+		<-ctx.Done()
+		shutdownCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+		defer cancel()
+		_ = server.Shutdown(shutdownCtx)
+	}()
+	_, _ = fmt.Fprintf(os.Stderr, "agentflow production API listening on %s scenario=%s queue=%s\n", listenURL(config.HTTPAddr), config.ScenarioFile, config.QueueKind)
+	if err := server.ListenAndServe(); err != nil && err != http.ErrServerClosed {
+		_, _ = fmt.Fprintln(os.Stderr, err)
+		os.Exit(1)
+	}
+}
+
+func listenURL(addr string) string {
+	host, port, err := net.SplitHostPort(addr)
+	if err != nil {
+		return "http://" + addr
+	}
+	if host == "" || host == "0.0.0.0" {
+		host = "localhost"
+	}
+	return "http://" + net.JoinHostPort(host, port)
+}