+++ cmd/agentctl/framework_helper.go
@@ -8,6 +8,14 @@ import (
 )
 
 func newFrameworkFromFlags(file, stateDir, tokenSecret string, tokenTTL time.Duration, tokenWriter io.Writer) (*agentflow.Framework, error) {
+	scenario, err := agentflow.LoadScenarioFile(file)
+	if err != nil {
+		return nil, err
+	}
+	workDir, err := agentflow.DemoWorkDir(file)
+	if err != nil {
+		return nil, err
+	}
 	opts := []agentflow.Option{
 		agentflow.WithHITLTokenSecret([]byte(tokenSecret), tokenWriter),
 		agentflow.WithHITLTokenTTL(tokenTTL),
@@ -19,5 +27,10 @@ func newFrameworkFromFlags(file, stateDir, tokenSecret string, tokenTTL time.Dur
 		}
 		opts = append(opts, agentflow.WithRunStateRepository(repo), agentflow.WithBlobStore(blobs))
 	}
-	return agentflow.NewFromFile(file, opts...)
+	demoOpts, err := agentflow.DemoOptions(scenario, agentflow.DemoConfig{WorkDir: workDir})
+	if err != nil {
+		return nil, err
+	}
+	opts = append(opts, demoOpts...)
+	return agentflow.New(scenario, opts...)
 }